Logout direkt in Menü eingebaut + Validierung bei Login

+++ server/src/routes/users.ts
@@ -1,35 +1,40 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import { AuthorisationService, JwtClaimSetHolder } from '../services/authorisationService';
+import validationService from '../services/validationService';
 import { User, UserInterface } from '../models/user';
 
 const router = Router();
 
 //Login
 router.post('/login', (request: Request, response: Response) => {
-const authorisationData = { email: request.body.email, password: request.body.password };
-const errors = [];
+    const authorisationData = { email: request.body.email, password: request.body.password };
+    const errors: string[] = [];
 
+    if (!validationService.hasRequiredFields(authorisationData, ['email', 'password'], errors)) {
+        response.status(400).json({ message: errors.join(' & ') });
+        return;
+    }
 
-	User.findOne({email: authorisationData.email})
-	.select('userId name email password')
-	.exec()
-	.then((user: UserInterface) => {
-		if (!user) {
-		return Promise.all([Promise.resolve(user), Promise.resolve(false)]);
+    User.findOne({ email: authorisationData.email })
+        .select('userId name email password')
+        .exec()
+        .then((user: UserInterface) => {
+            if (!user) {
+                return Promise.reject('User existiert nicht');
             } else {
                 return Promise.all([Promise.resolve(user), AuthorisationService.checkPassword(authorisationData.password, user)]);
             }
-	}).then(([user, isValid]: [UserInterface, boolean]) => {
+        }).then(([user, isValid]: [UserInterface, boolean]) => {
             if (isValid) {
                 return AuthorisationService.setTokenForUser(response, user);
             } else {
-                return Promise.reject('Email oder Password ungültig');
+                return Promise.reject('Password ist falsch');
             }
-	}).then(() => {
+        }).then(() => {
             response.sendStatus(201);
         }).catch((reason) => {
             AuthorisationService.removeToken(response);
-            response.status(400).json({message: reason});
+            response.status(400).json({ message: reason });
         });
 });
 
@@ -43,12 +48,12 @@ router.delete('/logout', (request: Request, response: Response) => {
 router.post('/registration', (request: Request, response: Response, next: NextFunction) => {
     const uD = request.body;
     var user = new User();
-    user.name = uD.name; 
+    user.name = uD.name;
     user.email = uD.email;
-    
+
 
     if (uD.password !== uD.password2) {
-        response.status(400).json({message: 'Password und Password wiederholen müssen gleich sein'});
+        response.status(400).json({ message: 'Password und Password wiederholen müssen gleich sein' });
     } else {
         User.findOne({ email: user.email }).exec().then((existingUser: UserInterface) => {
             if (existingUser) {
@@ -61,9 +66,9 @@ router.post('/registration', (request: Request, response: Response, next: NextFu
         }).then(() => {
             response.sendStatus(201);
         }).catch((reason: string) => {
-            response.status(400).json({message: reason});
+            response.status(400).json({ message: reason });
         });
-    } 
+    }
 });
 
 /*
@@ -73,7 +78,7 @@ router.post('/registration', (request: Request, response: Response, next: NextFu
 router.get('/getMe', (request: Request & JwtClaimSetHolder, response: Response) => {
     var user;
 
-    if (request.jwtClaimSet != null){
+    if (request.jwtClaimSet != null) {
         user = request.jwtClaimSet;
     }
